src/requirements.txt

aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.2.0
beautifulsoup4==4.13.3
black==24.10.0
cattrs==23.2.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
deprecated==1.2.18
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
ell-ai==0.0.16
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
joblib==1.4.2
lark==1.2.2
llama-cloud==0.1.14
llama-cloud-services==0.6.5
llama-index==0.12.22
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.23.post2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
loguru==0.7.2
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.65.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
propcache==0.3.0
psutil==5.9.8
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygithub==2.6.1
pygments==2.19.1
pyjwt==2.10.1
pymupdf==1.25.3
pymupdf4llm==0.0.17
pynacl==1.5.0
pypdf==5.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
semantic-text-splitter==0.24.1
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqlalchemy==2.0.39
sqlmodel==0.0.24
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
